+++ src/Local_Development/Base.php
@@ -257,6 +257,20 @@ class Base {
 		];
 		$file_data   = get_file_data( $filepath, $git_headers );
 
+		/**
+		 * Insert repositories added via GitHub Updater Additions plugin.
+		 *
+		 * @see GitHub Updater's Plugin or Theme class for definition.
+		 * @link https://github.com/afragen/github-updater-additions
+		 */
+		$additions = apply_filters( 'github_updater_additions', null, [], $type );
+		foreach ( (array) $additions as $slug => $headers ) {
+			if ( $slug === $file ) {
+				$file_data = array_merge( $file_data, $headers );
+				break;
+			}
+		}
+
 		foreach ( $file_data as $key => $value ) {
 			if ( ! empty( $value ) ) {
 				$githost = str_replace( "{$type_cap}URI", '', $key );