+++ CHANGES.md
@@ -3,6 +3,7 @@
 * add VCS checkouts automatically, thanks @Raruto
 * add Git Host icons as default setting for plugins and themes
 * automatically add localhost admin bar coloring when on localhost, default is on
+* add repositories that are added from [GitHub Update Additions](https://github.com/afragen/github-updater-additions)
 
 #### 2.4.1 / 2019-05-09
 * a11y updates for settings tabs